version update and minor refactoring
